+++ addon/components/charts/line-chart/component.js
@@ -2,9 +2,10 @@ import Component from '@ember/component';
 import layout from './template';
 import { isEmpty } from '@ember/utils';
 import { isArray } from '@ember/array';
+import { run } from '@ember/runloop';
 import { computed } from '@ember/object';
+import { htmlSafe } from '@ember/string';
 
-import SvgBase from 'building-blocks/mixins/svg-base';
 import { scaleLinear } from "d3-scale";
 import { axisLeft, axisBottom } from "d3-axis";
 import { select } from 'd3-selection';
@@ -15,12 +16,14 @@ const Y_AXIS_WIDTH = 35;
 const Y_TITLE_WIDTH = 20;
 const OFFSET = 15;
 
-export default Component.extend(SvgBase, {
+export default Component.extend({
   layout,
+  tagName: 'g',
   width: 500,
   height: 200,
+  dataTestId: 'line-chart',
 
-  isLoading: false,
+  isLoading: true,
   noDataAvailble: computed('data.[]', function() {
     let data = this.get('data');
     return isEmpty(data) || !isArray(data);
@@ -58,7 +61,16 @@ export default Component.extend(SvgBase, {
     let y = OFFSET;
     return `translate(${x}, ${y})`;
   }),
+  chartDimension: computed('width', 'height', function() {
+    let width = this.get('width') - Y_TITLE_WIDTH - Y_AXIS_WIDTH - OFFSET;
+    let height = this.get('height') - X_AXIS_HEIGHT - OFFSET;
+    return { width, height };
+  }),
 
+  didReceiveAttrs() {
+    this._super();
+    run.scheduleOnce('render', this, this.draw);
+  },
   draw() {
     if (this.get('noChart')) {
       return;
@@ -70,13 +82,13 @@ export default Component.extend(SvgBase, {
     let svgPlot = select(this.element);
     // y-axis
     let dataArrY = data.mapBy('y');
-    let yScale = scaleLinear().domain([Math.min(...dataArrY), Math.max(...dataArrY)]).range([this.get('height') - X_AXIS_HEIGHT - OFFSET, 0]);
+    let yScale = scaleLinear().domain([Math.min(...dataArrY), Math.max(...dataArrY)]).range([this.get('chartDimension.height'), 0]);
     let yAxisG = svgPlot.select('g.y-axis');
     let yAxis = axisLeft(yScale).ticks(6);
     yAxisG.call(yAxis);
     // x-axis
     let dataArrX = data.mapBy('x');
-    let xScale = scaleLinear().domain([Math.min(...dataArrX), Math.max(...dataArrX)]).range([0, this.get('width') - Y_TITLE_WIDTH - Y_AXIS_WIDTH - OFFSET]);
+    let xScale = scaleLinear().domain([Math.min(...dataArrX), Math.max(...dataArrX)]).range([0, this.get('chartDimension.width')]);
     let xAxisG = svgPlot.select('g.x-axis');
     let xAxis = axisBottom(xScale).ticks(6);
     xAxisG.call(xAxis);
@@ -86,5 +98,73 @@ export default Component.extend(SvgBase, {
     const yAccessor = (d) => yScale(d.y);
     let lineCurve = line().x(xAccessor).y(yAccessor); //.curve(curveLinear);
     linePath.datum(data).attr('d', lineCurve);
+  },
+
+  didInsertElement() {
+    this._super(...arguments);
+    this.addMouseEventHandlers();
+  },
+  willDestroyElement() {
+    this._super(...arguments);
+    this.removeMouseEventHandlers();
+  },
+
+  showTooltip: false,
+  addMouseEventHandlers() {
+    this.$('rect.svg-mouse-catch').on('mouseout', () => {
+      this.set('showTooltip', false);
+    });
+    this.$('rect.svg-mouse-catch').on('mouseover', (event) => {
+      if (this.get('noChart')) {
+        return;
+      }
+      this.handleMouseOver(event);
+    });
+    this.$('rect.svg-mouse-catch').on('mousemove', (event) => {
+      if (this.get('noChart')) {
+        return;
+      }
+      this.handleMouseOver(event);
+    });
+  },
+  handleMouseOver(event) {
+    let style;
+    let x = event.clientX;
+    let halfWindowWidth = window.innerWidth / 2;
+    if (x > halfWindowWidth) {
+      x = window.innerWidth - x;
+      style = `right: ${x}px; `;
+    } else {
+      style = `left: ${x}px; `;
+    }
+    let y = event.clientY;
+    let halfWindowHeight = window.innerHeight / 2;
+    if (y > halfWindowHeight) {
+      y = window.innerHeight - y;
+      style = `${style}bottom: ${y}px;`;
+    } else {
+      style = `${style}top: ${y}px;`;
+    }
+    this.set('tooltipStyle', new htmlSafe(style));
+
+    // mouse line
+    let offsetX = event.offsetX - (Y_TITLE_WIDTH + Y_AXIS_WIDTH + OFFSET);
+    let data = this.get('data');
+    let index = parseInt(Math.round(offsetX * data.get('length') / this.get('chartDimension.width'))) + 1;
+    if (index >=0 && index < data.length) {
+      this.set('dataValue', data[index].y);
+      this.set('offsetX', index * this.get('chartDimension.width') / (data.get('length') - 1));
+      this.set('showTooltip', true);
+    } else {
+      this.set('dataValue', null);
+      this.set('offsetX', null);
+      this.set('showTooltip', false);
+    }
+  },
+
+  removeMouseEventHandlers() {
+    this.$('rect.svg-mouse-catch').off('mouseout');
+    this.$('rect.svg-mouse-catch').off('mouseover');
+    this.$('rect.svg-mouse-catch').off('mousemove');
   }
 });