+++ addon/components/icons/facebook-icon/component.js
@@ -1,0 +1,7 @@
+import Component from '@ember/component';
+import layout from './template';
+import SvgBase from 'building-blocks/mixins/svg-base';
+
+export default Component.extend(SvgBase, {
+  layout
+});