+++ addon/components/icons/loading-icon/component.js
@@ -2,7 +2,6 @@ import Component from '@ember/component';
 import layout from './template';
 import SvgBase from 'building-blocks/mixins/svg-base';
 import { arc } from 'd3-shape';
-import { computed } from '@ember/object';
 import { task, timeout } from 'ember-concurrency';
 
 const DELAY_IN_MS = 50;
@@ -11,12 +10,6 @@ const ARC_RADIAN = 1.3;
 
 export default Component.extend(SvgBase, {
   layout,
-  status: "default",
-  width: computed.alias('widthHeight'),
-  height: computed.alias('widthHeight'),
-  radius: computed('widthHeight', function() {
-    return this.get('widthHeight') / 2;
-  }),
 
   draw() {
     this.get('_spin').perform();