+++ app/components/layout/main-section.js
@@ -1,0 +1,1 @@
+export { default } from 'building-blocks/components/layout/main-section/component';